+++ src/mkvtoolnix-gui/util/settings_names.h
@@ -109,6 +109,7 @@ char const * const s_valTrackInfo                           = "trackInfo";
 char const * const s_valType                                = "type";
 char const * const s_valUiDisableDarkStyleSheet             = "uiDisableDarkStyleSheet";
 char const * const s_valUiDisableHighDPIScaling             = "uiDisableHighDPIScaling";
+char const * const s_valUiDisableToolTips                   = "uiDisableToolTips";
 char const * const s_valUiFontFamily                        = "uiFontFamily";
 char const * const s_valUiFontPointSize                     = "uiFontPointSize";
 char const * const s_valUiLocale                            = "uiLocale";